+++ spec/after.js
@@ -1,2 +1,0 @@
-// put after Leaflet files as imagePath can't be detected in a PhantomJS env
-L.Icon.Default.imagePath = "/base/dist/images";