+++ spec/suites/core/EventsSpec.js
@@ -35,6 +35,62 @@ describe('Events', function () {
 			// expect(spy6.callCount).to.be(1);
 		});
 
+		it('fires all listeners in the order they are added', function () {
+			var obj = new L.Evented(),
+			    ctx1 = new L.Class(),
+			    ctx2 = new L.Class(),
+			    count = {one: 0, two: 0, three: 0, four: 0};
+
+			function listener1(e) {
+				count.one++;
+				expect(count.two).to.eql(0);
+			}
+
+			function listener2(e) {
+				count.two++;
+				expect(count.one).to.eql(1);
+				expect(count.three).to.eql(0);
+				if (count.two === 1) {
+					expect(this).to.eql(ctx2);
+				} else if (count.two === 2) {
+					expect(this).to.eql(ctx1);
+				} else {
+					expect(this).to.eql(obj);
+				}
+			}
+
+			function listener3(e) {
+				count.three++;
+				expect(count.two).to.eql(3);
+				expect(count.four).to.eql(0);
+				if (count.three === 1) {
+					expect(this).to.eql(ctx1);
+				} else if (count.three === 2) {
+					expect(this).to.eql(ctx2);
+				}
+			}
+
+			function listener4(e) {
+				count.four++;
+				expect(count.three).to.eql(2);
+			}
+
+			obj.on('test', listener1, ctx1);
+			obj.on('test', listener2, ctx2);
+			obj.on('test', listener2, ctx1);  // Same listener but with different context.
+			obj.on('test', listener2);  // Same listener but without context.
+			obj.on('test', listener3, ctx1);
+			obj.on('test', listener3, ctx2);
+			obj.on('test', listener4, ctx2);
+
+			obj.fireEvent('test');
+
+			expect(count.one).to.be(1);
+			expect(count.two).to.be(3);
+			expect(count.three).to.be(2);
+			expect(count.four).to.be(1);
+		});
+
 		it('provides event object to listeners and executes them in the right context', function () {
 			var obj = new L.Evented(),
 			    obj2 = new L.Evented(),
@@ -284,6 +340,44 @@ describe('Events', function () {
 			obj.fire('test');
 
 			expect(spy.called).to.be(false);
+		});
+
+		it('handles reentrant event firing', function () {
+			var obj = new L.Evented(),
+			    spy1 = sinon.spy(),
+			    spy2 = sinon.spy();
+
+			obj
+				.addEventListener('test1', function () {
+					obj.fire('test2');
+				})
+				.addEventListener('test2', spy1)
+				.addEventListener('test1', function () {
+					obj.removeEventListener('test1', spy2);
+				})
+				.addEventListener('test1', spy2);
+
+			obj.fireEvent('test1');
+			expect(spy1.called).to.be(true);
+			expect(spy2.called).to.be(false);
+		});
+
+		it('can remove an event listener while firing', function () {
+			var obj = new L.Evented(),
+			    spy = sinon.spy();
+
+			var removeSpy = function () {
+				obj.removeEventListener('test', spy);
+			};
+
+			obj.addEventListener('test', spy);
+			obj.addEventListener('test', removeSpy);
+
+			obj.fire('test');
+
+			obj.removeEventListener('test', removeSpy);
+
+			expect(obj.listens('test')).to.be(false);
 		});
 	});