+++ spec/suites/layer/TooltipSpec.js
@@ -43,6 +43,15 @@ describe('Tooltip', function () {
 		expect(map.hasLayer(layer._tooltip)).to.be(true);
 	});
 
+	it("can be added with bindTooltip before added to the map", function () {
+		var layer = new L.Marker(center);
+
+		layer.bindTooltip('Tooltip', {permanent: true});
+		expect(map.hasLayer(layer._tooltip)).to.be(false);
+		layer.addTo(map);
+		expect(map.hasLayer(layer._tooltip)).to.be(true);
+	});
+
 	it("is removed when removing marker", function () {
 		var layer = new L.Marker(center).addTo(map);
 
@@ -73,25 +82,63 @@ describe('Tooltip', function () {
 		expect(spy.calledOnce).to.be(true);
 	});
 
-	it("can be forced on top direction", function () {
+	it("honours offset on left direction", function () {
 		var layer = new L.Marker(center).addTo(map);
+		var spy = sinon.spy();
+		layer.on('click', spy);
+
+		layer.bindTooltip('A long tooltip that should be displayed on the left', {permanent: true, direction: 'left', interactive: true, offset: [-20, -20]});
+		expect(map.hasLayer(layer._tooltip)).to.be(true);
+		happen.at('click', 150, 180);
+		expect(spy.calledOnce).to.be(false);
+		happen.at('click', 130, 160);
+		expect(spy.calledOnce).to.be(true);
+	});
+
+	it("can be forced on top direction", function () {
+		var layer = L.circleMarker(center).addTo(map);
 		var spy = sinon.spy();
 		layer.on('click', spy);
 
 		layer.bindTooltip('A tooltip that should be displayed on the top', {permanent: true, direction: 'top', interactive: true});
 		expect(map.hasLayer(layer._tooltip)).to.be(true);
-		happen.at('click', 200, 170);  // Marker is on the map center, which is 400px large.
+		happen.at('click', 200, 180);  // Marker is on the map center, which is 400px large.
+		expect(spy.calledOnce).to.be(true);
+	});
+
+	it("honours offset on top direction", function () {
+		var layer = L.circleMarker(center).addTo(map);
+		var spy = sinon.spy();
+		layer.on('click', spy);
+
+		layer.bindTooltip('A tooltip that should be displayed on the top', {permanent: true, direction: 'top', interactive: true, offset: [-20, -20]});
+		happen.at('click', 200, 180);
+		expect(spy.calledOnce).to.be(false);
+		happen.at('click', 180, 150);
 		expect(spy.calledOnce).to.be(true);
 	});
 
 	it("can be forced on bottom direction", function () {
-		var layer = new L.Marker(center).addTo(map);
+		var layer = L.circleMarker(center).addTo(map);
 		var spy = sinon.spy();
 		layer.on('click', spy);
 
 		layer.bindTooltip('A tooltip that should be displayed on the top', {permanent: true, direction: 'bottom', interactive: true});
 		expect(map.hasLayer(layer._tooltip)).to.be(true);
-		happen.at('click', 200, 210);  // Marker is on the map center, which is 400px large.
+		happen.at('click', 200, 220);  // Marker is on the map center, which is 400px large.
+		expect(spy.calledOnce).to.be(true);
+	});
+
+	it("honours offset on bottom direction", function () {
+		var layer = L.circleMarker(center).addTo(map);
+		var spy = sinon.spy();
+		layer.on('click', spy);
+
+		layer.bindTooltip('A tooltip that should be displayed on the top', {permanent: true, direction: 'bottom', interactive: true, offset: [20, 20]});
+		expect(map.hasLayer(layer._tooltip)).to.be(true);
+		happen.at('click', 200, 220);
+		expect(spy.calledOnce).to.be(false);
+		happen.at('click', 220, 230);
 		expect(spy.calledOnce).to.be(true);
 	});
 
@@ -160,6 +207,19 @@ describe('Tooltip', function () {
 		expect(map.hasLayer(layer._tooltip)).to.be(true);
 	});
 
+	it("can be added on polygon with bindTooltip before beind added to the map", function () {
+		var layer = new L.Polygon([[55.8, 37.6], [55.9, 37.6], [55.8, 37.5]]);
+
+		layer.bindTooltip('Tooltip', {permanent: true});
+		expect(map.hasLayer(layer._tooltip)).to.be(false);
+		layer.addTo(map);
+		expect(map.hasLayer(layer._tooltip)).to.be(true);
+		layer.remove();
+		expect(map.hasLayer(layer._tooltip)).to.be(false);
+		layer.addTo(map);
+		expect(map.hasLayer(layer._tooltip)).to.be(true);
+	});
+
 	it("opens on polyline mouseover and close on mouseout", function () {
 		var layer = new L.Polyline([[55.8, 37.6], [55.9, 37.6], [55.8, 37.5]]).addTo(map);
 
@@ -173,11 +233,58 @@ describe('Tooltip', function () {
 		expect(map.hasLayer(layer._tooltip)).to.be(false);
 	});
 
-	it("stays open on marker when permanent", function () {
+	it("stays open on polyline when permanent", function () {
 		var layer = new L.Polyline([[55.8, 37.6], [55.9, 37.6], [55.8, 37.5]]).addTo(map);
 
 		layer.bindTooltip('Tooltip', {permanent: true});
 		expect(map.hasLayer(layer._tooltip)).to.be(true);
+	});
+
+	it("can be added on polyline with bindTooltip before added to the map", function () {
+		var layer = new L.Polyline([[55.8, 37.6], [55.9, 37.6], [55.8, 37.5]]);
+
+		layer.bindTooltip('Tooltip', {permanent: true});
+		expect(map.hasLayer(layer._tooltip)).to.be(false);
+		layer.addTo(map);
+		expect(map.hasLayer(layer._tooltip)).to.be(true);
+		layer.remove();
+		expect(map.hasLayer(layer._tooltip)).to.be(false);
+		layer.addTo(map);
+		expect(map.hasLayer(layer._tooltip)).to.be(true);
+	});
+
+	it("is opened when tapping on touch", function () {
+		var oldTouch = L.Browser.touch;
+		L.Browser.touch = true;
+		var layer = new L.Marker(center).addTo(map);
+
+		layer.bindTooltip('Tooltip');
+		expect(map.hasLayer(layer._tooltip)).to.be(false);
+		happen.click(layer._icon);
+		expect(map.hasLayer(layer._tooltip)).to.be(true);
+		L.Browser.touch = oldTouch;
+	});
+
+	it("is closed if not permanent when clicking on the map elsewhere on touch", function () {
+		var oldTouch = L.Browser.touch;
+		L.Browser.touch = true;
+		var layer = new L.Marker(center).addTo(map);
+
+		layer.bindTooltip('Tooltip');
+		happen.click(layer._icon);
+		expect(map.hasLayer(layer._tooltip)).to.be(true);
+		happen.click(map._container);
+		expect(map.hasLayer(layer._tooltip)).to.be(false);
+		L.Browser.touch = oldTouch;
+	});
+
+
+	it("opens with map.openTooltip", function (done) {
+		map.on('tooltipopen', function (e) {
+			expect(map.hasLayer(e.tooltip)).to.be(true);
+			done();
+		});
+		map.openTooltip('Tooltip', center);
 	});
 
 });