+++ src/core/Events.js
@@ -95,31 +95,20 @@ L.Evented = L.Class.extend({
 		/* get/init listeners for type */
 		var typeListeners = this._events[type];
 		if (!typeListeners) {
-			typeListeners = {
-				listeners: {},
-				count: 0
-			};
+			typeListeners = [];
 			this._events[type] = typeListeners;
 		}
 
-		var contextId = context && context !== this && L.stamp(context),
-		    newListener = {fn: fn, ctx: context};
-
-		if (!contextId) {
-			contextId = 'no_context';
-			newListener.ctx = undefined;
+		if (context === this) {
+			// Less memory footprint.
+			context = undefined;
 		}
-
-		// fn array for context
-		var listeners = typeListeners.listeners[contextId];
-		if (!listeners) {
-			listeners = [];
-			typeListeners.listeners[contextId] = listeners;
-		}
+		var newListener = {fn: fn, ctx: context},
+		    listeners = typeListeners;
 
 		// check if fn already there
 		for (var i = 0, len = listeners.length; i < len; i++) {
-			if (listeners[i].fn === fn) {
+			if (listeners[i].fn === fn && listeners[i].ctx === context) {
 				return;
 			}
 		}
@@ -129,68 +118,50 @@ L.Evented = L.Class.extend({
 	},
 
 	_off: function (type, fn, context) {
-		var typeListeners,
-		    contextId,
-		    listeners,
+		var listeners,
 		    i,
 		    len;
 
 		if (!this._events) { return; }
 
+		listeners = this._events[type];
+
+		if (!listeners) {
+			return;
+		}
+
 		if (!fn) {
 			// Set all removed listeners to noop so they are not called if remove happens in fire
-			typeListeners = this._events[type];
-			if (typeListeners) {
-				for (contextId in typeListeners.listeners) {
-					listeners = typeListeners.listeners[contextId];
-					for (i = 0, len = listeners.length; i < len; i++) {
-						listeners[i].fn = L.Util.falseFn;
-					}
-				}
-				// clear all listeners for a type if function isn't specified
-				delete this._events[type];
+			for (i = 0, len = listeners.length; i < len; i++) {
+				listeners[i].fn = L.Util.falseFn;
 			}
+			// clear all listeners for a type if function isn't specified
+			delete this._events[type];
 			return;
 		}
 
-		typeListeners = this._events[type];
-		if (!typeListeners) {
-			return;
+		if (context === this) {
+			context = undefined;
 		}
 
-		contextId = context && context !== this && L.stamp(context);
-		if (!contextId) {
-			contextId = 'no_context';
-		}
-
-		listeners = typeListeners.listeners[contextId];
 		if (listeners) {
 
 			// find fn and remove it
 			for (i = 0, len = listeners.length; i < len; i++) {
 				var l = listeners[i];
+				if (l.ctx !== context) { continue; }
 				if (l.fn === fn) {
 
 					// set the removed listener to noop so that's not called if remove happens in fire
 					l.fn = L.Util.falseFn;
-					typeListeners.count--;
 
-					if (len > 1) {
-						if (!this._isFiring) {
-							listeners.splice(i, 1);
-						} else {
-							/* copy array in case events are being fired */
-							typeListeners.listeners[contextId] = listeners.slice();
-							typeListeners.listeners[contextId].splice(i, 1);
-						}
-					} else {
-						delete typeListeners.listeners[contextId];
+					if (this._firingCount) {
+						/* copy array in case events are being fired */
+						this._events[type] = listeners = listeners.slice();
 					}
+					listeners.splice(i, 1);
 
 					return;
-				}
-				if (listeners.length === 0) {
-					delete typeListeners.listeners[contextId];
 				}
 			}
 		}
@@ -206,23 +177,16 @@ L.Evented = L.Class.extend({
 		var event = L.Util.extend({}, data, {type: type, target: this});
 
 		if (this._events) {
-			var typeListeners = this._events[type];
+			var listeners = this._events[type];
 
-			if (typeListeners) {
-				this._isFiring = true;
-
-				// each context
-				for (var contextId in typeListeners.listeners) {
-					var listeners = typeListeners.listeners[contextId];
-
-					// each fn in context
-					for (var i = 0, len = listeners.length; i < len; i++) {
-						var l = listeners[i];
-						l.fn.call(l.ctx || this, event);
-					}
+			if (listeners) {
+				this._firingCount = (this._firingCount + 1) || 1;
+				for (var i = 0, len = listeners.length; i < len; i++) {
+					var l = listeners[i];
+					l.fn.call(l.ctx || this, event);
 				}
 
-				this._isFiring = false;
+				this._firingCount--;
 			}
 		}
 
@@ -237,8 +201,8 @@ L.Evented = L.Class.extend({
 	// @method listens(type: String): Boolean
 	// Returns `true` if a particular event type has any listeners attached to it.
 	listens: function (type, propagate) {
-		var typeListeners = this._events && this._events[type];
-		if (typeListeners && typeListeners.count) { return true; }
+		var listeners = this._events && this._events[type];
+		if (listeners && listeners.length) { return true; }
 
 		if (propagate) {
 			// also check parents for listeners if event propagates